fix(array_list): correct to_string memory deallocation error
Use the correct argument `(void**) &element_string` instead of `(void**) element_string`. `memory_free` expects a pointer to a pointer, but a raw pointer was passed instead.

This issue was hidden by the mocked `free` function in tests, which has now been removed, alongside with the `array_list_delete` test, which is unnecessary.